refactor(cli): use gqlURL from @oasis-sh/shared
