+++ packages/cli/src/commands/search.ts
@@ -1,11 +1,11 @@
-import * as log from '../utils/log';
+import * as log from '@oasis-sh/shared';
+import { gqlURL } from '@oasis-sh/shared';
 import { gql, GraphQLClient } from 'graphql-request';
-import { GQL_URL } from '../constants';
 
 export async function handler(yargs: any) {
   const useJSON = yargs.json ?? false;
 
-  const client = new GraphQLClient(GQL_URL);
+  const client = new GraphQLClient(gqlURL);
 
   const searchQuery = yargs._.slice(1).join(' ');
   const limit = yargs.limit ?? 10.0;
