--- packages/cli/src/constants.ts
@@ -1,1 +1,0 @@
-export const GQL_URL = 'https://dev.oasis.sh/graphql';
