+++ packages/shared/src/lib/constants.ts
@@ -1,0 +1,1 @@
+export const gqlURL = 'https://dev.oasis.sh/graphql';
