+++ packages/shared/src/output/exit.ts
