+++ packages/shared/src/output/log.ts
@@ -3,6 +3,8 @@ import chalk from 'chalk';
 export const colors = {
   error: chalk.red('error') + ' -',
   ready: chalk.green('ready') + ' -',
+  warn: chalk.yellow('warn') + ' -',
+  event: chalk.magenta('event') + ' -',
   info: chalk.magenta('info') + ' -',
 };
 
@@ -14,6 +16,14 @@ export const ready = (...message: string[]) => {
 // Uh oh, there were issues found
 export const error = (...message: string[]) => {
   console.error(colors.error, ...message);
+};
+
+export const warn = (...message: string[]) => {
+  console.warn(colors.warn, ...message);
+};
+
+export const event = (...message: string[]) => {
+  console.log(colors.event, ...message);
 };
 
 // Information
